pyproject.toml
[project]
name = "week3-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.136.0",
    "google-genai>=1.43.0",
    "kagglehub>=1.0.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "pydantic-settings>=2.11.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "uvicorn>=0.38.0",
    "xgboost>=3.2.0",
    "google>=3.0.0",
    "voyageai>=0.3.7",
]
